requirements.txt
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
bcrypt==3.2.0
boto3==1.20.46
botocore==1.23.46
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
databases==0.5.5
decorator==5.1.1
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.73.0
fuzzywuzzy==0.18.0
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
jmespath==0.10.0
makefun==1.13.1
Mako==1.1.6
MarkupSafe==2.0.1
outcome==1.1.0
passlib==1.7.4
Pillow==9.0.0
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
PyJWT==1.7.1
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
python-Levenshtein==0.12.2
python-multipart==0.0.5
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
s3transfer==0.5.0
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
starlette==0.17.1
transliterate==1.10.2
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.0.post1
uvloop==0.16.0
validators==0.18.2
watchgod==0.7
webdriver-manager==3.5.3
websockets==10.1
wsproto==1.0.0
